Check for existence

+++ index.js
@@ -5,7 +5,12 @@ const fs = require('fs-extra');
 module.exports = {
   name: 'ember-cli-netlify',
   outputReady(result) {
-    fs.copySync('.netlifyheaders', 'dist/_headers', { clobber: true });
-    fs.copySync('.netlifyredirects', 'dist/_redirects', { clobber: true });
+    if (fs.pathExistsSync('.netlifyheaders')) {
+      fs.copySync('.netlifyheaders', 'dist/_headers', { clobber: true });
+    }
+
+    if (fs.pathExistsSync('.netlifyredirects')) {
+      fs.copySync('.netlifyredirects', 'dist/_redirects', { clobber: true });
+    }
   }
 };
